SCOTCH: Last resort to compile PT-SCOTCH.
On some platforms (Theta, Lengau, etc), a compiled binary (even
serial) must be run with the MPI job launcher.  The standard way
to compile PT-SCOTCH will fail in such cases.  This commit added
an option "SCOTCH_LAST_RESORT" which generates header files of
PT-SCOTCH by launching the compiled binary with the MPI launcher.
This, together with requesting an interactive job or submitting
the compilation as a job to the queuing system, may successfully
compile PT-SCOTCH.  If this still doesn't work, then separately
compiled SuperLU_DIST and PT-SCOTCH libraries are the true last
resort.