Test: Increased tolerance for PEXSI.
Since we only test two dummy SCF steps, the results computed by
PEXSI would depend on the number of shifts in the inertia counting
phase, which is determined by the number of MPI tasks.  This leads
to slightly different results for different numbers of MPI tasks.
In real settings the effect of the initial counting phase always
fades away as the chemical potential converges together with the
SCF cycle.