cmake: Replace -std=.. flags with compile features
Instead of requiring the user to supply the -std=c99 and -std=c++11
flags, define the c_std_99 and cxx_std_11 compile features for targets
that require them. This is portable and overall should be the
preferred approach.

I have declared the compile features as PRIVATE as I cannot think of a
reason for PUBLIC at the moment. This can be changed in the future.