CMake: Fix multiple target bug for imported ELSI
When ELSI is imported as an external CMake project, if the importing
project passes in libraries to ELSI, ELSI will create a new target
for the libraries.  This may trigger a "target already defined"
error if ELSI chooses the same name when it creates its target as the
importing project did as part of its build system.

We now tell ELSI to simply ignore creating the target if it already
exists.  This fix implicitly assumes that the importing code isn't doing
anything really stupid (like giving its LAPACK libraries a target name
of ScaLAPACK, for example) and thus is not ideal.