[CMake] Fix search path for external libraries
When using the find_library command, CMake will search for libraries in
the PATHS variables *after* it has searched system folders.  This will
lead to unclear errors when the user is unable to specify different
versions of a library.  Using the HINTS variable instead allows the user
to override system folders, yielding the expected behavior.