Clean up useless solver decision code
The current implementation needs too much memory. When trying the ELPA,
PEXSI, and NTPoly solvers one after another, the interface needs to
keep the memory allocation of all the three solvers, which would run
out of memory quickly and would make this feature useless for all but
very small test systems. A different approach is needed here.

The "quick decision" layer is not touched and should still work.